refactor: combines all card components into one
