+++ less/components/consonant/cards-grid.less
@@ -16,7 +16,7 @@
         justify-content: center;
         width: 100%;
         max-width: 100%;
-        grid-template-columns: repeat(auto-fit, minmax(@consonantCardMinWidth, max-content));
+        grid-template-columns: repeat(auto-fit, minmax(@consonantCardMinWidth, 1fr));
 
         /* Spacing variations */
 
@@ -43,18 +43,12 @@
         /* Cards variations */
 
         &--2up {
-            grid-template-columns: repeat(auto-fit, minmax(@consonant-DoubleWideCard-mobileMinWidth, max-content));
+            grid-template-columns: repeat(auto-fit, minmax(@consonant-DoubleWideCard-mobileMinWidth, 1fr));
 
-            .consonant-ThreeFourthCard,
-            .consonant-OneHalfCard,
-            .consonant-HalfHeightCard,
-            .consonant-DoubleWideCard,
-            .consonant-ProductCard,
-            .consonant-TextCard,
-            .consonant-FullCard {
+            .consonant-Card {
                 max-width: @consonant-DoubleWideCard-maxWidth;
 
-                &-inner {
+                &-content {
                     width: @consonant-DoubleWideCard-maxWidth;
                     max-width: 100%;
                 }
@@ -62,13 +56,8 @@
         }
 
         &--3up {
-            .consonant-ThreeFourthCard,
-            .consonant-OneHalfCard,
-            .consonant-HalfHeightCard,
-            .consonant-ProductCard,
-            .consonant-TextCard,
-            .consonant-FullCard {
-                &-inner {
+            .consonant-Card {
+                &-content {
                     width: @consonantCardMaxWidth;
                     max-width: 100%;
                 }
@@ -76,145 +65,68 @@
         }
 
         &--4up {
-            grid-template-columns: repeat(auto-fit, minmax(@consonant-CardsGrid-smallerCardMinWidth, max-content));
+            grid-template-columns: repeat(auto-fit, minmax(@consonant-CardsGrid-smallerCardMinWidth, 1fr));
 
-            .consonant-ThreeFourthCard,
-            .consonant-OneHalfCard,
-            .consonant-HalfHeightCard,
-            .consonant-ProductCard,
-            .consonant-TextCard,
-            .consonant-FullCard {
+            .consonant-Card {
                 min-width: @consonant-CardsGrid-smallerCardMinWidth;
                 max-width: @consonant-CardsGrid-smallerCardMaxWidth;
                 height: @consonant-CardsGrid-smallerCardheight;
 
-                &-inner {
+                &-content {
                     width: @consonant-CardsGrid-smallerCardMaxWidth;
                     max-width: 100%;
                 }
             }
 
-            .consonant-ThreeFourthCard .consonant-ThreeFourthCard-img {
-                min-height: @consonant-CardsGrid-smallerCardImgThreeFourthHeight;
-            }
-
-            .consonant-ThreeFourthCard > a > .consonant-ThreeFourthCard-text {
-                display: none;
-            }
-
-            .consonant-ThreeFourthCard > a > .consonant-ThreeFourthCard-title {
-                margin-bottom: 0;
-                max-height: 1.375rem * 2;
-
-                .line-clamp(2);
-
-                &:first-child {
-                    max-height: 1.375rem * 3;
-
-                    .line-clamp(3);
-                }
-            }
-
-            .consonant-ProductCard,
-            .consonant-TextCard,
-            .consonant-OneHalfCard {
+            &.one-half .consonant-Card {
                 height: auto;
-            }
 
-            .consonant-OneHalfCard .consonant-OneHalfCard-img {
-                min-height: @consonant-CardsGrid-smallerCardImgOneHalfHeight;
-                max-height: @consonant-CardsGrid-smallerCardImgOneHalfHeight;
-            }
-
-            .consonant-HalfHeightCard {
-                height: @consonant-OneHalfCard-height;
+                .consonant-Card-header {
+                    min-height: @consonant-CardsGrid-smallerCardImgOneHalfHeight;
+                    max-height: @consonant-CardsGrid-smallerCardImgOneHalfHeight;
+                }
             }
         }
 
         &--5up {
-            grid-template-columns: repeat(auto-fit, minmax(@consonant-CardsGrid-smallestCardMinWidth, max-content));
+            grid-template-columns: repeat(auto-fit, minmax(@consonant-CardsGrid-smallestCardMinWidth, 1fr));
 
-            .consonant-ThreeFourthCard,
-            .consonant-OneHalfCard,
-            .consonant-HalfHeightCard,
-            .consonant-ProductCard,
-            .consonant-TextCard,
-            .consonant-FullCard {
+            .consonant-Card {
                 min-width: @consonant-CardsGrid-smallestCardMinWidth;
                 max-width: @consonant-CardsGrid-smallestCardMaxWidth;
                 height: @consonant-CardsGrid-smallestCardHeight;
 
-                &-inner {
+                &-content {
                     width: @consonant-CardsGrid-smallestCardMaxWidth;
                     max-width: 100%;
                 }
             }
 
-            .consonant-ThreeFourthCard .consonant-ThreeFourthCard-img {
-                min-height: @consonant-CardsGrid-smallestCardImgThreeFourthHeight;
-            }
-
-            .consonant-ThreeFourthCard > a > .consonant-ThreeFourthCard-text {
-                display: none;
-            }
-
-            .consonant-ThreeFourthCard > a > .consonant-ThreeFourthCard-title {
-                margin-bottom: 0;
-                max-height: 1.375rem * 2;
-
-                .line-clamp(2);
-
-                &:first-child {
-                    max-height: 1.375rem * 3;
-
-                    .line-clamp(3);
-                }
-            }
-
-            .consonant-ProductCard,
-            .consonant-TextCard,
-            .consonant-OneHalfCard {
+            &.one-half .consonant-Card {
                 height: auto;
-            }
 
-            .consonant-OneHalfCard .consonant-OneHalfCard-img {
-                min-height: @consonant-CardsGrid-smallestCardImgOneHalfHeight;
-                max-height: @consonant-CardsGrid-smallestCardImgOneHalfHeight;
-            }
-
-            .consonant-HalfHeightCard {
-                height: @consonant-OneHalfCard-height;
-            }
-        }
-
-        &--3up&--doubleWideCards,
-        &--4up&--doubleWideCards,
-        &--5up&--doubleWideCards {
-            grid-template-columns: repeat(auto-fit, minmax(@consonant-DoubleWideCard-mobileMinWidth, max-content));
-
-            .consonant-DoubleWideCard-inner {
-                width: @consonant-DoubleWideCard-maxWidth;
-                max-width: 100%;
+                .consonant-Card-header {
+                    min-height: @consonant-CardsGrid-smallestCardImgOneHalfHeight;
+                    max-height: @consonant-CardsGrid-smallestCardImgOneHalfHeight;
+                }
             }
         }
 
         @media @consonant-680px-up {
             &--2up {
-                grid-template-columns: repeat(auto-fit, minmax(@consonant-DoubleWideCard-desktopMinWidth, max-content));
+                grid-template-columns: repeat(auto-fit, minmax(@consonant-DoubleWideCard-desktopMinWidth, 1fr));
 
-                .consonant-ThreeFourthCard,
-                .consonant-OneHalfCard,
-                .consonant-HalfHeightCard,
-                .consonant-DoubleWideCard,
-                .consonant-FullCard {
+                .consonant-Card {
                     min-width: @consonant-DoubleWideCard-desktopMinWidth;
                 }
             }
 
-            &--3up&--doubleWideCards,
-            &--4up&--doubleWideCards,
-            &--5up&--doubleWideCards {
-                grid-template-columns: repeat(auto-fit, minmax(@consonant-DoubleWideCard-desktopMinWidth, max-content));
+            &--3up,
+            &--4up,
+            &--5up {
+                &.double-wide {
+                    grid-template-columns: repeat(auto-fit, minmax(@consonant-DoubleWideCard-desktopMinWidth, 1fr));
+                }
             }
         }
     }
