+++ less/components/consonant/main.less
@@ -29,6 +29,7 @@
 @import 'filters/left/chosen-item.less';
 
 /* Cards; */
+@import 'cards/card.less';
 @import 'cards/one-half.less';
 @import 'cards/three-fourth.less';
 @import 'cards/full.less';
