+++ less/components/consonant/theme.less
@@ -283,7 +283,7 @@
                     }
                 }
 
-                &-img {
+                &-header {
                     background-color: @consonantGray300;
                 }
 
@@ -317,7 +317,7 @@
                     background-color: @consonantBaseColor;
                 }
 
-                &-img {
+                &-header {
                     background-color: @consonantGray300;
                 }
 
@@ -355,7 +355,7 @@
                     background-color: @consonantBaseColor;
                 }
 
-                &-img {
+                &-header {
                     background-color: @consonantGray300;
                 }
 
@@ -430,7 +430,7 @@
                     }
                 }
 
-                &-img {
+                &-header {
                     background-color: @consonantGray300;
                 }
