+++ react/src/js/components/Consonant/Cards/__tests__/Full.spec.js
@@ -1,7 +1,7 @@
 import { screen } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
 
-import Card from '../Full';
+import Card from '../Card';
 
 import { DEFAULT_PROPS_FULL } from '../../Testing/Constants/Card';
 
@@ -22,9 +22,11 @@ describe('Consonant/Card/Full-Card', () => {
                     },
                 },
             },
-        } = renderCard();
+        } = renderCard({
+            cardStyle: 'full-card',
+        });
 
-        const bannerElement = screen.getByTestId('consonant-FullCard-banner');
+        const bannerElement = screen.getByTestId('consonant-Card-banner');
         const bannerIconElement = screen.getByTestId('consonant-Card-bannerImg');
 
         expect(bannerElement).toHaveStyle({
@@ -43,13 +45,17 @@ describe('Consonant/Card/Full-Card', () => {
                     },
                 },
             },
-        } = renderCard();
+        } = renderCard({
+            cardStyle: 'full-card',
+        });
 
         const badgeElement = screen.queryByText(someBadgeText);
         expect(badgeElement).not.toBeNull();
     });
     test('should be able to render a logo overlay', () => {
-        renderCard();
+        renderCard({
+            cardStyle: 'full-card',
+        });
         const logoAltText = screen.getByAltText('logo-alt-text');
         expect(logoAltText).not.toBeNull();
     });
