+++ react/src/js/components/Consonant/Container/__tests__/ResultsPerPage.spec.js
@@ -31,7 +31,7 @@ describe('Consonant/Container/Results Per Page', () => {
         await act(async () => render(<Container config={configToUse} />));
         await waitFor(() => screen.getByTestId('consonant-CardsGrid'));
 
-        expect(screen.queryAllByTestId('consonant-OneHalfCard')).toHaveLength(9);
+        expect(screen.queryAllByTestId('consonant-Card')).toHaveLength(9);
     });
 
     test('should render 1 card if authored to do so', async () => {
@@ -41,7 +41,7 @@ describe('Consonant/Container/Results Per Page', () => {
         await act(async () => render(<Container config={configToUse} />));
         await waitFor(() => screen.getByTestId('consonant-CardsGrid'));
 
-        expect(screen.queryAllByTestId('consonant-OneHalfCard')).toHaveLength(1);
+        expect(screen.queryAllByTestId('consonant-Card')).toHaveLength(1);
     });
 
     test('should render all cards if limit > total cards count', async () => {
@@ -52,6 +52,6 @@ describe('Consonant/Container/Results Per Page', () => {
         await waitFor(() => screen.getByTestId('consonant-CardsGrid'));
 
         const allCards = cards.length;
-        expect(screen.queryAllByTestId('consonant-OneHalfCard')).toHaveLength(allCards);
+        expect(screen.queryAllByTestId('consonant-Card')).toHaveLength(allCards);
     });
 });
